+++ foxes/utils/__init__.py
@@ -7,7 +7,7 @@ from .pandas_utils import PandasFileHelper
 from .xarray_utils import write_nc
 from .subclasses import all_subclasses
 from .dict import Dict
-from .factory import Factory, FDict
+from .factory import Factory, FDict, WakeKFactory
 from .data_book import DataBook
 from .cubic_roots import cubic_roots
 from .geopandas_utils import read_shp, shp2csv, read_shp_polygons, shp2geom2d