+++ foxes/utils/factory.py
@@ -1,5 +1,3 @@
-import numpy as np
-
 from .dict import Dict
 
 
@@ -73,14 +71,10 @@ class Factory:
         self.var2arg = var2arg
         self.hints = hints
 
-        parts = name_template.split(">")
-        if len(parts) < 2:
-            raise ValueError(
-                f"Factory '{name_template}': Expecting at least one variable in template, pattern '<..>'"
-            )
-
         self._vars = []
         self._pre = []
+        parts = name_template.split(">")
+
         for i, p in enumerate(parts):
             if i < len(parts) - 1:
                 parts2 = p.split("<")
@@ -262,6 +256,106 @@ class Factory:
         return self.base(*self.args, **kwargs)
 
 
+class WakeKFactory:
+    """
+    A factory that automatically handles
+    wake_k parameters
+
+    Attributes
+    ----------
+    factories: list of Factory
+        The individual factories
+
+    :group: utils
+
+    """
+    def __init__(self, base, name_template, *args, hints={}, **kwargs):
+        """
+        Constructor.
+        
+        Parameters
+        ----------
+        base: class
+            The class of which objects are to be created
+        name_template: str
+            The name template, e.g. 'name_<A>_<B>_<C>' for
+            variables A, B, C. Indicate wake_k part by '_[wake_k]'
+        args: tuple, optional
+            Additional arguments for Factory
+        hints: dict
+            Hints for print_toc, only for variables for which the
+            options are functions or missing
+        kwargs: dict
+            Additional arguments for Factory
+
+        """
+        self._base = base
+        self._kwargs = kwargs
+        self._template0 = name_template
+        self.factories = []
+        i0 = name_template.find("_[wake_k]")
+        i1 = i0 + len("_[wake_k]")
+
+        if i0 < 0:
+            raise ValueError(f"String '_[wake_k]' not found in name template '{name_template}'")
+        else:
+            # add case ka, kb:
+            t = name_template[:i0] + "_ka<ka>_kb<kb>"
+            if len(name_template) > i1:
+                t += name_template[i1:]
+            h = hints.copy()
+            h["ka"] = "(Value, e.g. 04 for 0.4)"
+            h["kb"] = "(Value, e.g. 001 for 0.01)"
+            self.factories.append(Factory(
+                base, t, *args, hints=h, **kwargs,
+                ka=lambda x: float(f"0.{x[1:]}" if x[0] == "0" else float(x)),
+                kb=lambda x: float(f"0.{x[1:]}" if x[0] == "0" else float(x)),
+            ))
+
+            # add case ka:
+            t = name_template[:i0] + "_ka<ka>"
+            if len(name_template) > i1:
+                t += name_template[i1:]
+            h = hints.copy()
+            h["ka"] = "(Value, e.g. 04 for 0.4)"
+            self.factories.append(Factory(
+                base, t, *args, hints=h, **kwargs,
+                ka=lambda x: float(f"0.{x[1:]}" if x[0] == "0" else float(x)),
+            ))
+
+            # add case k:
+            t = name_template[:i0] + "_k<k>"
+            if len(name_template) > i1:
+                t += name_template[i1:]
+            h = hints.copy()
+            h["k"] = "(Value, e.g. 004 for 0.04)"
+            self.factories.append(Factory(
+                base, t, *args, hints=h, **kwargs,
+                k=lambda x: float(f"0.{x[1:]}" if x[0] == "0" else float(x)),
+            ))
+
+        # add case without k:
+        t = name_template[:i0]
+        if len(name_template) > i1:
+            t += name_template[i1:]
+        self.factories.append(Factory(
+            base, t, *args, hints=hints, **kwargs,
+        ))
+
+    def __str__(self):
+        """String representation"""
+        s = f"{self._template0}: {self._base.__name__} with"
+        for k, d in self._kwargs.items():
+            s += f"\n  {k}={d}"
+        f0 = self.factories[-1]
+        for v in f0.variables:
+            if v in f0.options and isinstance(f0.options[v], dict):
+                s += f"\n  {v} from {list(f0.options[v])}"
+            else:
+                s += f"\n  {v}={f0.hints.get(v, '(value)')}"
+        s += f"\n  [wake_k]=(None or k<k> or ka<ka> or ka<ka>_kb<kb>, e.g. 004 for 0.04)"
+        return s
+
 class FDict(Dict):
     """
     A dictionary with factory support
@@ -295,19 +389,21 @@ class FDict(Dict):
         self.store_created = store_created
         self.factories = []
 
-    def add_factory(self, *args, **kwargs):
+    def add_factory(self, *args, factory=None, **kwargs):
         """
-        Constructor.
+        Adds a Factory object.
 
         Parameters
         ----------
         args: tuple, optional
             Parameters for the Factory constructor
+        factory: Factory, optional
+            The factory object
         kwargs: dict, optional
             Parameters for the Factory constructor
 
         """
-        f = Factory(*args, **kwargs)
+        f = Factory(*args, **kwargs) if factory is None else factory
         i = len(self.factories)
         for gi in range(len(self.factories) - 1, -1, -1):
             g = self.factories[gi]
@@ -322,6 +418,21 @@ class FDict(Dict):
             self.factories.append(f)
         else:
             self.factories.insert(i, f)
+
+    def add_k_factory(self, *args, **kwargs):
+        """
+        Adds a WakeKFactory.
+
+        Parameters
+        ----------
+        args: tuple, optional
+            Parameters for the Factory constructor
+        kwargs: dict, optional
+            Parameters for the Factory constructor
+
+        """
+        for f in WakeKFactory(*args, **kwargs).factories:
+            self.add_factory(factory=f)
 
     def __contains__(self, key):
         found = super().__contains__(key)